+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/rodekruis/IBF-system/compare/0.1.0...master)
 
 ### Added
+- 2021-02-05: Covid data needs to be uploaded to '/uga-data-level-2/upload'
 - 2021-01-27: Add WATERPOINTDATA_TOKEN to .env file
+
 ### Removed
 
 ### Changed